app/image-resizer-appv1.0/app/zip-processing/zip-processed-listener/requirements.txt
confluent-kafka
sqlalchemy
psycopg2-binary
python-dotenv
